Task 1. List length

+++ 0x17-doubly_linked_lists/1-dlistint_len.c
@@ -8,7 +8,7 @@
 
 size_t dlistint_len(const dlistint_t *h)
 {
-	size_t elementCount = 0;
+	int elementCount = 0;
 	const dlistint_t *temp = h;
 
 	while (temp != NULL)
